corrigido validações e concluido metodos de crud

+++ src/main/java/br/com/Academy/Calisthenics/service/AlunoService.java
@@ -1,11 +1,10 @@
 package br.com.Academy.Calisthenics.service;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.Academy.Calisthenics.entity.Aluno;
+import br.com.Academy.Calisthenics.exception.AplicationException;
 import br.com.Academy.Calisthenics.repository.AlunoRepository;
 
 @Service
@@ -16,21 +15,42 @@ public class AlunoService {
 	
 	
 	public Aluno salvarAluno(Aluno aluno) {
+		
+		verificarAluno(aluno.getCpf());
+		
 		return alunoRepository.save(aluno);
 	}
 	
+	public void verificarAluno(String cpf) {
+		if(alunoRepository.existsBydCpf(cpf)) {
+			throw new AplicationException("Aluno ja Existe!");
+		}
+	}
+	
 	public Aluno buscarAluno(Long id) {
-		Optional<Aluno> aluno = alunoRepository.findById(id);
-		return aluno.get();
+		Aluno aluno = alunoRepository.getOne(id);
+		if(aluno != null) {
+			return aluno;
+		}
+		return null;
 	}
 	
 	public void excluirAluno(Long id) {
-		alunoRepository.deleteById(id);
+		Aluno aluno = buscarAluno(id);
+		if(aluno != null) {
+			alunoRepository.deleteById(id);
+		}
+		throw new AplicationException("Usuario Invalido!");
+		
 	}
 	
-	public Aluno atualizarDadosAluno(Aluno aluno, Long id) {
-		aluno.setId(id);
-		Aluno novoAluno = salvarAluno(aluno);
-		return novoAluno;
+	public void atualizarDadosAluno(Aluno aluno, Long id) {
+		aluno = buscarAluno(id);
+		if(aluno != null) {
+			Aluno novoAluno = salvarAluno(aluno);
+			
+		}
+		throw new AplicationException("");
+		
 	}
 }
